refactor(integer): refactor from/to_be/le_bytes
There was a bug in to_be_bytes_slice, it was missing a
`slice.reverse()` (from_be_bytes correctly has it)

The test value used did not catch that so we change the value
used to expose the problem.

And the from/to functions have been refactored to used
from_be_bytes / to_be_bytes, etc from the stdlib to
only have one layer of endianess to manage.